修复bug

+++ android/app/src/main/java/com/xtkj/paopaoxiche/bean/IncomeBean.java
@@ -47,7 +47,7 @@ public class IncomeBean {
          */
 
         private long time;
-        private int totalMoney;
+        private double totalMoney;
         private List<ItemsBean> items;
 
         public long getTime() {
@@ -58,11 +58,11 @@ public class IncomeBean {
             this.time = time;
         }
 
-        public int getTotalMoney() {
+        public double getTotalMoney() {
             return totalMoney;
         }
 
-        public void setTotalMoney(int totalMoney) {
+        public void setTotalMoney(double totalMoney) {
             this.totalMoney = totalMoney;
         }
 
@@ -86,7 +86,7 @@ public class IncomeBean {
             private int carType;
             private String carTypeText;
             private int id;
-            private int money;
+            private double money;
             private String title;
 
             public int getCarType() {
@@ -113,11 +113,11 @@ public class IncomeBean {
                 this.id = id;
             }
 
-            public int getMoney() {
+            public double getMoney() {
                 return money;
             }
 
-            public void setMoney(int money) {
+            public void setMoney(double money) {
                 this.money = money;
             }
